dev: new wallet connect btn
